+++ api/observer_test.go
@@ -2,7 +2,7 @@ package api
 
 import (
 	"github.com/stretchr/testify/assert"
-	"github.com/trustwallet/blockatlas/observer"
+	"github.com/trustwallet/blockatlas/observer/storage"
 	"testing"
 )
 
@@ -21,15 +21,15 @@ func Test_parseSubscriptions(t *testing.T) {
 			},
 			webhook: "http://127.0.0.1:8080",
 			wantSubs: []interface{}{
-				&observer.Subscription{
+				&storage.Subscription{
 					Coin: uint(2), Webhook: "http://127.0.0.1:8080",
 					Address: "zpub6rH4MwgyTmuexAX6HAraks5cKv5BbtmwdLirvnU5845ovUJb4abgjt9DtXK4ZEaToRrNj8dQznuLC6Nka4eMviGMinCVMUxKLpuyddcG9Vc",
 				},
-				&observer.Subscription{
+				&storage.Subscription{
 					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6BpYi6J1GZzfY3yY7DbhLLccF3efQa18nQngM3jaehgtNSoEgk6UtPULpC3oK5oA3trczY8Ld34LFw1USMPfGHwTEizdD5QyGcMyuh2UoBA",
 				},
-				&observer.Subscription{
+				&storage.Subscription{
 					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6CYwPfnPJLPquufPkb98coSb3mdy1CgaZrWUtYWGJTJ4VWZUbzH9HLGy7nHpP7DG4UdTkYYpirkTWQSP7pWHsrk24Nos5oYNHpfr4BgPVTL",
 				},
@@ -41,11 +41,11 @@ func Test_parseSubscriptions(t *testing.T) {
 			},
 			webhook: "http://127.0.0.1:8080",
 			wantSubs: []interface{}{
-				&observer.Subscription{
+				&storage.Subscription{
 					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6BpYi6J1GZzfY3yY7DbhLLccF3efQa18nQngM3jaehgtNSoEgk6UtPULpC3oK5oA3trczY8Ld34LFw1USMPfGHwTEizdD5QyGcMyuh2UoBA",
 				},
-				&observer.Subscription{
+				&storage.Subscription{
 					Coin: uint(0), Webhook: "http://127.0.0.1:8080",
 					Address: "xpub6CYwPfnPJLPquufPkb98coSb3mdy1CgaZrWUtYWGJTJ4VWZUbzH9HLGy7nHpP7DG4UdTkYYpirkTWQSP7pWHsrk24Nos5oYNHpfr4BgPVTL",
 				},
